==> scripts/02_clean_hive_data.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Purpose

This script processes data are from experiments using arrays of artificial flowers conducted in the summer of 2022. Specifically, it cleans and reorganizes data from records of bees returning to the hive after visiting artificial flower arrays on six mornings (2 trials). It uses output from 01_clean_array_data to add information about which array each bee had visited and to add bees that visited the arrays but never danced (and thus perhaps we have no data about those bees in the hive spreadsheets).


# Set up working directory

```{r, setup} 
library(knitr) 
opts_knit$set(root.dir = normalizePath('../'))
```


# Load libraries

```{r}
library(tidyverse)
```


# Read in hive and flower array csv files

_Note:_ The "_VideoAnalysis2.csv" files are files from the first Trial in which we reanalyzed the hive videos, and this time we used the array visit information to determine when we should expect bees to return to the hive to better capture all dances.

```{r}
hive_0615 <- read.csv("data/0615_VideoAnalysis_2.csv")
hive_0616 <- read.csv("data/0616_VideoAnalysis_2.csv")
hive_0617 <- read.csv("data/0617_VideoAnalysis_2.csv")

hive_0718 <- read.csv("data/0718_VideoAnalysis.csv", skip = 1)
hive_0719 <- read.csv("data/0719_VideoAnalysis.csv", skip = 1)
hive_0720 <- read.csv("data/0720_VideoAnalysis.csv", skip = 1) %>%
  filter(Bee != "")  # gets rid of extra blank lines

# Output from 01_clean_array_data.Rmd
array_df <- read.csv("cleaned_data/01_combined_flower_array_data.csv")
```

Make a list of these data frames to process more efficiently (the for loops that iterate over this list are based on code from this Stack Overflow answer https://stackoverflow.com/a/62216494):

```{r}
hive_dfs <- list(hive_0615,
                      hive_0616,
                      hive_0617,
                      hive_0718,
                      hive_0719,
                      hive_0720
               )
names(hive_dfs) <- c("hive_0615",
                      "hive_0616",
                      "hive_0617",
                      "hive_0718",
                      "hive_0719",
                      "hive_0720"
               )
```


# Select columns we will use

These include:

* Colony
* Date
* Bee
* Entrance minute
* Entrance second
* Danced
* Bouts columns (1-14)

_Note:_ Bouts are not really meaningful. They are just groupings of waggle runs that were convenient while recording data and do not necessarily represent sets of waggle runs with a certain duration gap in between them. We just sum them to get the Total_runs for each dance (or 0 for visits without a dance).

```{r}
select_hive_columns <- . %>%
  select(Colony, 
         Date, 
         Bee, 
         Entrance_minute,
         Entrance_second,
         Bout_1,
         Bout_2,
         Bout_3,
         Bout_4,
         Bout_5,
         Bout_6,
         Bout_7,
         Bout_8,
         Bout_9,
         Bout_10,
         Bout_11,
         Bout_12,
         Bout_13,
         Bout_14)
```

```{r}
for (i in 1:length(hive_dfs)){
  hive_dfs[[i]] <- select_hive_columns(hive_dfs[[i]])
  list2env(hive_dfs,.GlobalEnv)
}
```


# Step : Combine hive data frames into one

Now that we have just those columns, we'll combine the data frames.

```{r}
combined_hive_df <- rbind(hive_dfs[[1]], hive_dfs[[2]])
for (i in 3:length(hive_dfs)){
  combined_hive_df <- rbind(combined_hive_df, hive_dfs[[i]])
}
```


# Step : Convert bee IDs to standard format

Use periods instead of dashes to match the flower array ID format and add Colony to the bee ID to deal with repeated color combinations used across trials of the experiment

```{r}
for (row in 1:length(combined_hive_df$Bee)){
    
  combined_hive_df$Bee[row] <- gsub(pattern = "-",
                        replacement = ".",
                        x = combined_hive_df$Bee[row])
}

# Add colony ID to bee ID (bee "BO" from colony "7B" -> "BO_7B")
combined_hive_df <- combined_hive_df %>%
  mutate(Bee = paste(Bee, Colony, sep = "_"))

```


# Step : Calculate decimal entrance time and total waggle runs

Then we need to calculate the decimal time of each trip entrance in the hive and add the a column for the sum of the waggle runs as well:

```{r}
combined_hive_df <- combined_hive_df %>%
  mutate(Decimal_entrance_time = Entrance_minute + (Entrance_second/60)) %>%
  # for some reason, the dance bout columns were loaded in as logical vectors so we need to convert them to integers before summing them
  mutate(Bout_1 = as.integer(Bout_1)) %>% 
  mutate(Bout_2 = as.integer(Bout_2)) %>% 
  mutate(Bout_3 = as.integer(Bout_3)) %>% 
  mutate(Bout_4 = as.integer(Bout_4)) %>% 
  mutate(Bout_5 = as.integer(Bout_5)) %>% 
  mutate(Bout_6 = as.integer(Bout_6)) %>% 
  mutate(Bout_7 = as.integer(Bout_7)) %>% 
  mutate(Bout_8 = as.integer(Bout_8)) %>% 
  mutate(Bout_9 = as.integer(Bout_9)) %>% 
  mutate(Bout_10 = as.integer(Bout_10)) %>% 
  mutate(Bout_11 = as.integer(Bout_11)) %>% 
  mutate(Bout_12 = as.integer(Bout_12)) %>% 
  mutate(Bout_13 = as.integer(Bout_13)) %>% 
  mutate(Bout_14 = as.integer(Bout_14)) %>% 
  mutate(Total_runs = rowSums(across(
                          c(Bout_1,
                          Bout_2,
                          Bout_3,
                          Bout_4,
                          Bout_5,
                          Bout_6,
                          Bout_7,
                          Bout_8,
                          Bout_9,
                          Bout_10,
                          Bout_11,
                          Bout_12,
                          Bout_13,
                          Bout_14)
                          ),
                          na.rm = T)) %>%
  mutate(Danced = ifelse(Total_runs > 0, 1, 0)) %>%
  # and now we don't need all those Bout columns
  select(!starts_with("Bout"))
```


# Step : Fix Bee IDs that differed in flower array records and remove problematic bees

There were unfortunately some problems created by using paints with so many similar colors and one bee appeared to visit both the array and surrounding flowers (for pollen) so the code below corrects problems that can be corrected and removes bees for which we cannot be sure of the array that they visited.

```{r}
combined_hive_df <- combined_hive_df %>%
  # OP.a is called OF.a at the feeder and that really is more
  # consistent with her appearance
  mutate(Bee = ifelse(Bee == "OP.a_7B", "OF.a_7B", Bee)) %>%
  # Unfortunately, GY and GP seem to be indistinguishable
  # I think it was always GY that was dancing, but both clearly visited and
  # their marks are not consistently different
  filter(Bee != "GP_7B") %>%
  filter(Bee != "GY_7B") %>%
  # Also, BWW is difficult to distinguish from BW 
  # and it's possible that she is the same bee who may have visited
  # both arrays at different times (mostly tremble danced)
  filter(Bee != "BWW_7B") %>%
  # WR was seen with pollen on her legs after visiting the feeder
  # so I think she may have been visiting multiple sources
  filter(Bee != "WR_7B")
```


# Step : Combine hive data with flower array data

A full join allows us to assign information about the array each bee was visiting and the treatment on that day as well as to add in bees that never danced (and thus should be recorded as performing 0 dances for later analysis).

```{r}
combined_hive_df <- combined_hive_df %>%
  full_join(array_df)
```

Then we can check that the bees that visited the hive but weren't recorded at a flower array make sense:

```{r}
combined_hive_df %>%
  filter(is.na(Array_time)) %>%
  group_by(Date, Bee) %>%
  summarize(Dances = sum(Danced))
```
On 6-15-22:

* _O-a may not have visited the feeder and she didn't dance

On 7-18-22:

* RB-a appears to enter the hive once but then wanders around and follows a few dances- she may not have visited any food source
* WG appears to enter the hive once but was not recorded at the array and didn't dance

On 7-20-22: 

* GWW entered the hive once. She didn't dance, and I assume she was visiting something else (the extra white may even have been pollen)

So we can remove them:

```{r}
combined_hive_df <- combined_hive_df %>%
  filter(!is.na(Array_time))
```

Now we can check how many dances were earlier than the first time we saw that bee at the array:

```{r}
combined_hive_df %>%
  filter(Danced == 1) %>%
  filter(Decimal_entrance_time < Array_time) %>%
  select(Date, Bee, Decimal_entrance_time, Array_time)
```

Based on this, I think it makes sense to give some leeway of 5 minutes (approximately one trip) in trial 2 because it was hard to keep track of every flower array visit.

```{r}
combined_hive_df <- combined_hive_df %>%
  filter(is.na(Decimal_entrance_time) | (Decimal_entrance_time - Array_time > -5)) # either the bee was never recorded in the hive or she was recorded at a reasonable time
```


# Step : Create "cleaned_data" folder if needed

```{r}
if (!dir.exists("cleaned_data")){
  dir.create("cleaned_data")
}
```


# Step : Write data frame of dances to a csv file

All recorded hive visits:

```{r}
combined_hive_df %>%
  filter(!is.na(Danced)) %>%  # either 0 or 1 in the Danced column means it is a hive record
  write.csv(.,
          "cleaned_data/02_all_hive_visits.csv",
          row.names = F)
```

All recorded hive visits that included at least one waggle run:

```{r}
combined_hive_df %>%
  filter(Danced == 1) %>%
  write.csv(., 
          "cleaned_data/02_all_dances.csv", 
          row.names = F)
```


# Step : Make a summary by bee and date (and include all treatment info)

```{r}
hive_summary <- combined_hive_df %>%
  group_by(Colony, Trial, Date, Day, Scent, Treatment, Setup, Site, Density, Flower_number, Bee) %>%
  summarize(Dances = sum(Danced, na.rm = T), 
            Waggle_runs = sum(Total_runs, na.rm = T)) %>%
  # if the bee was added because she appears in the array records but not the hive records
  # we need to make the following columns 0 for later analyses
  mutate(Dances = ifelse(is.na(Dances), 0, Dances)) %>%
  mutate(Waggle_runs = ifelse(is.na(Waggle_runs), 0, Waggle_runs))
```

Write the summary to a csv file:

```{r}
write.csv(hive_summary,
          "cleaned_data/02_per_bee_summary.csv",
          row.names = F)
```
