==> scripts/03_regression_analyses.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Purpose

This script uses generalized linear mixed models to examine the effects of density and number of flowers on the number/rate of dances that bees performed and the number of waggle runs performed per dance. The data are from experiments using arrays of artificial flowers conducted in the summer of 2022.


# Set up working directory

```{r, setup} 
library(knitr) 
opts_knit$set(root.dir = normalizePath('../'))
```


# Load libraries

```{r}
library("tidyverse")  # for wrangling data
library("lme4")       # for glmer.nb function
library("nlme")       # for lme function
library("glmmTMB")    # for fitting generalized mixed models
library("bbmle") # for making delta AIC comparison table
library("DHARMa")     # to assess generalized mixed models
library("sjPlot")     # to get table of incidence rate ratios
```


# Read in data files

```{r}
# This file has one row per dance
dances_df <- read.csv("cleaned_data/02_all_dances.csv") %>%
  mutate(ObsID = row_number()) %>%
  # We'll make sure that results indicate increases with higher density and
  # number of flowers so that they are easier to interpret
  mutate(Density = factor(Density, levels = c("Low", "High"))) %>%
  mutate(Flower_number = factor(Flower_number, levels = c("Low", "High")))

# This file has one row per uniquely-IDed bee
bees_df <- read.csv("cleaned_data/02_per_bee_summary.csv") %>%
  mutate(ObsID = row_number()) %>%
  # Again, we'll make sure that results indicate increases with higher density and
  # number of flowers
  mutate(Density = factor(Density, levels = c("Low", "High"))) %>%
  mutate(Flower_number = factor(Flower_number, levels = c("Low", "High")))

```


# Summary to check data files read in correctly

```{r}
dances_summary <- bees_df %>%
  group_by(Trial, Date, Site) %>%
  summarize(Bees = n_distinct(Bee), Dances = sum(Dances))

dances_summary
```

```{r}
waggle_runs_summary <- dances_df %>%
  group_by(Trial, Date, Site) %>%
  summarize(Dancing_bees = n_distinct(Bee), Mean_runs = mean(Total_runs), Sum_runs = sum(Total_runs))

waggle_runs_summary
```

```{r}
full_summary <- full_join(dances_summary, waggle_runs_summary) %>%
  mutate(Prop_danced = Dancing_bees / Bees)

full_summary
```


# Fitting a set of generalized mixed models

```{r}
# poisson with observation-level random effect
model_dance_olrs <- glmmTMB(Dances ~ Density + Flower_number +
                          (1|Scent) + (1|Date) + (1|Trial) + (1|Bee) + (1|ObsID),
                        family = "poisson", 
                        data = bees_df)

# negative binomial with linear parameterization
model_dance_nb1 <- glmmTMB(Dances ~ Density + Flower_number + 
                         (1|Scent) + (1|Date) + (1|Trial) + (1|Bee), 
                         family = "nbinom1", 
                         data = bees_df)

# negative binomial with quadratic parameterization
model_dance_nb2 <- glmmTMB(Dances ~ Density + Flower_number + 
                          (1|Scent) + (1|Date) + (1|Trial) +  (1|Bee), 
                         family="nbinom2", 
                         data = bees_df)

# zero-inflated model with constant zero probability
model_dance_zip <- glmmTMB(Dances ~ Density + Flower_number +
                          (1|Scent) + (1|Date) + (1|Trial) + (1|Bee),  
                          zi = ~1,
                          family="poisson", 
                          data = bees_df)

# zero-inflated model with zero probability dependent on same factors as number of dances
model_dance_zip2 <- glmmTMB(Dances ~ Density + Flower_number + 
                          (1|Scent) + (1|Date) + (1|Trial) + (1|Bee),  
                          zi = ~.,
                          family="poisson", 
                          data = bees_df)
```

```{r}
AICtab(model_dance_olrs,
       model_dance_nb1,
       model_dance_nb2,
       model_dance_zip,
       model_dance_zip2)
```

```{r}
summary(model_dance_olrs)
```

```{r}
simulationOutput_olrs <- simulateResiduals(
  fittedModel = model_dance_olrs, plot = FALSE)

plot(simulationOutput_olrs)
```


```{r}
summary(model_dance_nb1)
```

```{r}
simulationOutput_nb1 <- simulateResiduals(
  fittedModel = model_dance_nb1, plot = FALSE)

plot(simulationOutput_nb1)
```

## Now we can get a table of the incidence rate ratios

I believe that in this case we can interpret the incidence rate ratio as the relative probably (comparing 2 treatments) that a bee visiting the feeder will decide to perform a waggle dance within the observation period.


```{r}
sjPlot::tab_model(model_dance_nb1)
```



```{r}
summary(model_dance_nb2)
```

```{r}
simulationOutput_nb2 <- simulateResiduals(
  fittedModel = model_dance_nb2, plot = FALSE)

plot(simulationOutput_nb2)
```

```{r}
summary(model_dance_zip)
```

```{r}
simulationOutput_zip <- simulateResiduals(
  fittedModel = model_dance_zip, plot = FALSE)

plot(simulationOutput_zip)
```

```{r}
summary(model_dance_zip2)
```

```{r}
simulationOutput_zip2 <- simulateResiduals(
  fittedModel = model_dance_zip2, plot = FALSE)

plot(simulationOutput_zip2)
```


# Waggle runs/dance model comparison

```{r}
model_runs_olrs <- glmmTMB(Total_runs ~ Density + Flower_number + 
                          (1|Scent) + (1|Date) + (1|Trial) + (1|Bee) + (1|ObsID),
                        family = "poisson", 
                        data = dances_df)

model_runs_nb1 <- glmmTMB(Total_runs ~ Density + Flower_number + 
                          (1|Scent) + (1|Date) + (1|Trial) + (1|Bee), 
                         family = "nbinom1", 
                         data = dances_df)

model_runs_nb2 <- glmmTMB(Total_runs ~ Density + Flower_number + 
                          (1|Scent) + (1|Date) + (1|Trial) +  (1|Bee), 
                         family="nbinom2", 
                         data = dances_df)
```

```{r}
AICtab(model_runs_olrs,
       model_runs_nb1,
       model_runs_nb2)
```

```{r}
summary(model_runs_olrs)
```

```{r}
simulationOutput_runs_olrs <- simulateResiduals(
  fittedModel = model_runs_olrs, plot = FALSE)

plot(simulationOutput_runs_olrs)
```

```{r}
summary(model_runs_nb1)
```

```{r}
simulationOutput_runs_nb1 <- simulateResiduals(
  fittedModel = model_runs_nb1, plot = FALSE)

plot(simulationOutput_runs_nb1)
```

```{r}
summary(model_runs_nb2)
```

```{r}
simulationOutput_runs_nb2 <- simulateResiduals(
  fittedModel = model_runs_nb2, plot = FALSE)

plot(simulationOutput_runs_nb2)
```

## Now we can get a table of the incidence rate ratios

I believe that in this case we can interpret the incidence rate ratio as the relative probably (comparing 2 treatments) that, once a bee has finished a waggle run, she will decide to perform another waggle run within the same hive visit.

```{r}
sjPlot::tab_model(model_runs_nb2)
```
