==> scripts/04_boxplots.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Purpose

This script creates the main figures showing the effects of flower patch treatments on individual bees' dancing and the waggle runs performed per dance. The data are from experiments using arrays of artificial flowers conducted in the summer of 2022.


# Set up working directory

```{r, setup} 
library(knitr) 
opts_knit$set(root.dir = normalizePath('../'))
```


# Load libraries

```{r}
library("tidyverse")
library("RColorBrewer")    # for colorblind-friendly color palettes
```


# Read in data files

```{r}
# This file has one row per dance
dances_df <- read.csv("cleaned_data/02_all_dances.csv") %>%
  mutate(Day = as.factor(Day))

# This file has one row per uniquely-IDed bee
bees_df <- read.csv("cleaned_data/02_per_bee_summary.csv") %>%
  mutate(Day = as.factor(Day))

```


# Create "figures" folder if needed

```{r}
if (!dir.exists("figures")){
  dir.create("figures")
}
```


# Make boxplot figures

Figure 2A: Dances per bee

```{r}
bees_df %>%
  ggplot(aes(x = Day, y = Dances, fill = Setup)) +
  geom_boxplot() +
  scale_fill_brewer(palette = "Paired") +
  labs(y = "Dances performed per bee") +
  facet_grid(Trial~Treatment, scales = "free_y") +
  theme_classic(base_size = 16)

ggsave("figures/04_boxplot_dances_per_bee.jpg", 
       device="jpeg",
       width = 9.5,
       height = 7,
       units = "in",
       dpi = 300)
```

Figure 2B: Waggle runs per dance

```{r}
dances_df %>%
  ggplot(aes(x = Day, y = Total_runs, fill = Setup)) +
  geom_boxplot() +
  scale_fill_brewer(palette = "Paired") +
  labs(y = "Waggle runs per dance") +
  facet_grid(Trial~Treatment, scales = "free_y")+
  theme_classic(base_size = 16)

ggsave("figures/04_boxplot_waggle_runs_per_dance.jpg", 
       device="jpeg",
       width = 9.5,
       height = 7,
       units = "in",
       dpi = 300)
```
