==> scripts/05_cumulative_figures.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Purpose

This script creates the cumulative figures to show how density and number of flowers affected the accumulation of bees visiting the arrays, bees engaged in dancing, dances, and waggle runs at the colony level. The data are from experiments using arrays of artificial flowers conducted in the summer of 2022.


# Set up working directory

```{r, setup} 
library(knitr) 
opts_knit$set(root.dir = normalizePath('../'))
```


# Load libraries

```{r}
library("tidyverse")
library("RColorBrewer")    # for colorblind-friendly color palettes
```


# Read in data files

```{r}
# This file has one row per hive visit
dances_df <- read.csv("cleaned_data/02_all_dances.csv") %>%
  mutate(Day = as.factor(Day))  # this was read in as an integer vector

# This file has one row per array visit
array_visits_df <- read.csv("cleaned_data/01_all_array_visits.csv") %>%
  mutate(Day = as.factor(Day))  # this was read in as an integer vector
```


# Add a column so we can get cumulative totals per minute of each trial

```{r}
dances_df <- dances_df %>%
  mutate(Integer_time = floor(Decimal_entrance_time))
```


# Calculate cumulative bees, dances, and waggle runs

```{r}
# Adding cumulative bees (I think!) requires a set of nested for loops

## We need to create one list of bees for every trial, day, treatment combination
## We use these lists to create a smaller data frame with cumulative bees added
## and then we need to add each of these temporary data frames to a larger
## data frame

# Here's where the larger data frame is initialized
cumulative_df <- data.frame(Trial = character(),
                 Day = character(),
                 Treatment = character(),
                 Density = character(),
                 Flower_number = character(),
                 Minute = integer(),
                 Cumul_bees = integer(),
                 Cumul_dances = integer(),
                 Cumul_waggle_runs = integer())

# Here are all of the options to loop through
trials <- unique(dances_df$Trial)
days <- unique(dances_df$Day)
treatments <- unique(dances_df$Treatment)

# Now we start the loops
for (trial_i in trials){
  # Trial 1-> 2 hours, Trial 2-> 1 hour
  if (trial_i == "Trial 1"){
    minutes <- 1:119
  } else {
    minutes <- 1:59
  }
  for (day_i in days){
    for (treatment_i in treatments){
      # get that subset of rows
      temp_df <- dances_df %>%
        filter(Trial == trial_i & Day == day_i & Treatment == treatment_i)
      # get the treatment values to add to the temporary data frame
      curr_density <- temp_df$Density[1]
      curr_flower_number <- temp_df$Flower_number[1]
      curr_setup <- temp_df$Setup[1]
      # create a temporary df to store the cumulative values
      temp_cumul_df <- data.frame(Trial = rep(x = trial_i, times = length(minutes)),
                                  Day = rep(x = day_i, times = length(minutes)),
                                  Treatment = rep(x = treatment_i, times = length(minutes)),
                                  Density = rep(x = curr_density, times = length(minutes)),
                                  Flower_number = rep(x = curr_flower_number, times = length(minutes)),
                                  Setup = rep(x = curr_setup, times = length(minutes)),
                                  Minute = minutes,
                                  Cumul_bees = rep(x = 0, times = length(minutes)),
                                  Cumul_dances = rep(x = 0, times = length(minutes)),
                                  Cumul_waggle_runs = rep(x = 0, times = length(minutes)))
      for (row in 1:length(minutes)){
        # save the minute for the current row
        curr_minute <- row
        # get a subset of temp_df
        temp_df_row <- temp_df %>%
          filter(Integer_time <= curr_minute)
        # get a vector of all bees that entered during/before that minute
        curr_bees_list <- unique(temp_df_row$Bee)
        # add the cumulative totals to the temporary data frame
        temp_cumul_df$Cumul_bees[row] <- length(curr_bees_list)
        temp_cumul_df$Cumul_dances[row] <- sum(temp_df_row$Danced)
        temp_cumul_df$Cumul_waggle_runs[row] <- sum(temp_df_row$Total_runs)
      }
      # then add the temporary df to the df we initialized above
      cumulative_df <- rbind(cumulative_df, temp_cumul_df)
    }
  }
}
```


# Calculate cumulative bees, dances, and waggle runs

This is a very similar process to the one above for dances, but with the data on visits by bees to the flower arrays instead.

```{r}
## We need to create one list of bees for every trial, day, treatment combination
## We use these lists to create a smaller data frame with cumulative bees added
## and then we need to add each of these temporary data frames to a larger
## data frame

# Here's where the larger data frame is initialized
cumulative_array_df <- data.frame(Trial = character(),
                 Day = character(),
                 Treatment = character(),
                 Density = character(),
                 Flower_number = character(),
                 Minute = integer(),
                 Cumul_array_bees = integer())

# Here are all of the options to loop through
trials <- unique(dances_df$Trial)
days <- unique(dances_df$Day)
treatments <- unique(dances_df$Treatment)

# Now we start the loops
for (trial_i in trials){
  # Trial 1-> 2 hours, Trial 2-> 1 hour
  if (trial_i == "Trial 1"){
    minutes <- 1:119
  } else {
    minutes <- 1:59
  }
  for (day_i in days){
    for (treatment_i in treatments){
      # get that subset of rows
      temp_df <- array_visits_df %>%
        filter(Trial == trial_i & Day == day_i & Treatment == treatment_i)
      # get the treatment values to add to the temporary data frame
      curr_density <- temp_df$Density[1]
      curr_flower_number <- temp_df$Flower_number[1]
      curr_setup <- temp_df$Setup[1]
      # create a temporary df to store the cumulative values
      temp_cumul_df <- data.frame(Trial = rep(x = trial_i, times = length(minutes)),
                                  Day = rep(x = day_i, times = length(minutes)),
                                  Treatment = rep(x = treatment_i, times = length(minutes)),
                                  Density = rep(x = curr_density, times = length(minutes)),
                                  Flower_number = rep(x = curr_flower_number, times = length(minutes)),
                                  Setup = rep(x = curr_setup, times = length(minutes)),
                                  Minute = minutes,
                                  Cumul_array_bees = rep(x = 0, times = length(minutes)))
      for (row in 1:length(minutes)){
        # save the minute for the current row
        curr_minute <- row
        # get a subset of temp_df
        temp_df_row <- temp_df %>%
          filter(Array_time <= curr_minute)
        # get a vector of all bees that were seen during/before that minute
        curr_bees_list <- unique(temp_df_row$Bee)
        # add the cumulative totals to the temporary data frame
        temp_cumul_df$Cumul_array_bees[row] <- length(curr_bees_list)
      }
      # then add the temporary df to the df we initialized above
      cumulative_array_df <- rbind(cumulative_array_df, temp_cumul_df)
    }
  }
}
```


# Combine cumulative_df and cumulative_array_df so we can scale by number of bees that
# have visited the array

```{r}
combined_df <- full_join(cumulative_df, cumulative_array_df) %>%
  # scale bees that danced by number that visited feeder
  mutate(Scaled_dancing_bees = ifelse(Cumul_array_bees > 0, Cumul_bees / Cumul_array_bees, 0),
         # scale number of dances by number of bees that visited feeder
         Scaled_dances = ifelse(Cumul_array_bees > 0, Cumul_dances / Cumul_array_bees, 0),
         # scale number of waggle runs by number of bees that visited feeder
         Scaled_waggle_runs = ifelse(Cumul_array_bees > 0, Cumul_waggle_runs / Cumul_array_bees, 0))
```


# Create "figures" folder if needed

```{r}
if (!dir.exists("figures")){
  dir.create("figures")
}
```


# Making cumulative figures

Unique bees who danced

```{r}
cumulative_df %>%
  ggplot(aes(x = Minute, y = Cumul_bees, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative bees that danced") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)

ggsave("figures/05_lineplot_cumulative_dancing_bees.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
  
```

Figure S2: Unique bees who danced scaled by the cumulative number of bees who had visited the feeder

```{r}
combined_df %>%
  ggplot(aes(x = Minute, y = Scaled_dancing_bees, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative proportion of bees that visited the array that also danced") +
  facet_grid(Trial~Treatment) +  # trials had very different totals
  theme_light(base_size = 18)

ggsave("figures/05_lineplot_scaled_cumulative_dancing_bees.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
  
```

Dances performed

```{r}
cumulative_df %>%
  ggplot(aes(x = Minute, y = Cumul_dances, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative dances") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)


ggsave("figures/05_lineplot_cumulative_dances.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
```

Figure 3A: Dances performed scaled by the cumulative number of bees that had visited the array

```{r}
combined_df %>%
  ggplot(aes(x = Minute, y = Scaled_dances, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative dances per bee that visited array") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)


ggsave("figures/05_lineplot_scaled_cumulative_dances.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
```

Waggle runs performed

```{r}
cumulative_df %>%
  ggplot(aes(x = Minute, y = Cumul_waggle_runs, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative waggle runs") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)
  

ggsave("figures/05_lineplot_cumulative_waggle_runs.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
```

Figure 3B: Waggle runs performed scaled by the cumulative number of bees that had visited the array

```{r}
combined_df %>%
  ggplot(aes(x = Minute, y = Scaled_waggle_runs, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative waggle runs per bee that visited array") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)
  

ggsave("figures/05_lineplot_scaled_cumulative_waggle_runs.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
```

Figure S3: Unique bees who visited flower array

This shows how many bees could possibly have decided to contribute to the number of dances and waggle runs.

```{r}
cumulative_array_df %>%
  ggplot(aes(x = Minute, y = Cumul_array_bees, color = Setup, linetype = Day)) +
  scale_linetype_manual(values = c("twodash", "solid", "dotted")) +
  scale_color_brewer(palette = "Paired") +
  geom_line(linewidth = 1) +
  labs(y = "Cumulative bees that visted flower array") +
  facet_grid(Trial~Treatment, scales = "free_y") +  # trials had very different totals
  theme_light(base_size = 18)

ggsave("figures/05_lineplot_cumulative_array_visiting_bees.jpg", 
       device="jpeg",
       width = 10.5,
       height = 7,
       units = "in",
       dpi = 300)
  
```
